refactor(mint-burn): refactor duplicated code
