+++ packages/block-indexer/docs/register_token.md
@@ -62,3 +62,62 @@ Transaction: e {
   "blockHeight": "0x13b",
   "txIndex": "0x0"
 }
+
+``` json
+Transaction result of register non-token on Moonbeam
+
+{
+  "blockHash": "0x7af144284932676713c3504b967e58624e748e6e49fd2dbe664d9886c3916280",
+  "blockNumber": 203560,
+  "chainId": "0x501",
+  "creates": null,
+  "from": "0xf24FF3a9CF04c71Dbc94D0b566f7A27B94566cac",
+  "gas": 323804,
+  "gasPrice": "1000000000",
+  "hash": "0xf0640b5877203550332354046e6ed56f9cd881ed6fa4a30b9e39f4be62ae3256",
+  "input": "0xf2c298be0000000000000000000000000000000000000000000000000000000000000020000000000000000000000000000000000000000000000000000000000000000453414e4700000000000000000000000000000000000000000000000000000000",
+  "nonce": 23,
+  "publicKey": "0x509540919faacf9ab52146c9aa40db68172d83777250b28e4679176e49ccdd9fa213197dc0666e85529d6c9dda579c1295d61c417f01505765481e89a4016f02",
+  "r": "0xec0addb75b2cf805d76fb10b23ad9262e02c9947aa35d65a919ff01f6f442625",
+  "raw": "0xf8cb17843b9aca008304f0dc947d4567b7257cf869b01a47e8cf0edb3814bdb96380b864f2c298be0000000000000000000000000000000000000000000000000000000000000020000000000000000000000000000000000000000000000000000000000000000453414e4700000000000000000000000000000000000000000000000000000000820a26a0ec0addb75b2cf805d76fb10b23ad9262e02c9947aa35d65a919ff01f6f442625a01848b5ace46973bdd44f0222435135967d833128bf8c14b477dd7ac2a87636e5",
+  "s": "0x1848b5ace46973bdd44f0222435135967d833128bf8c14b477dd7ac2a87636e5",
+  "standardV": "0x1",
+  "to": "0x7d4567B7257cf869B01a47E8cf0EDB3814bDb963",
+  "transactionIndex": 2,
+  "v": "0xa26",
+  "value": "0",
+  "receipt": {
+    "blockHash": "0x7af144284932676713c3504b967e58624e748e6e49fd2dbe664d9886c3916280",
+    "blockNumber": 203560,
+    "contractAddress": null,
+    "cumulativeGasUsed": 3044719,
+    "from": "0xf24ff3a9cf04c71dbc94d0b566f7a27b94566cac",
+    "gasUsed": 70609,
+    "logs": [],
+    "logsBloom": "0x00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000",
+    "status": true,
+    "to": "0x7d4567b7257cf869b01a47e8cf0edb3814bdb963",
+    "transactionHash": "0xf0640b5877203550332354046e6ed56f9cd881ed6fa4a30b9e39f4be62ae3256",
+    "transactionIndex": 2
+  }
+}
+```
+
+# Register non-token on Moonbeam
+### Get list of token already registed
+
+```bash
+eth contract:call --network http://localhost:9933 bshcore@0x7d4567B7257cf869B01a47E8cf0EDB3814bDb963 "coinNames()"
+```
+
+### Register non-token
+
+```bash
+eth contract:send --pk <addmin_private_key>  --network http://localhost:9933 bshcore@<address_bsh_core> "register('<token_name>')"
+```
+
+### Check transaction result
+
+```bash
+eth transaction:get --network http://localhost:9933 <transaction_hash>
+```
